refactor test location and re-run pre-commit
